+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Quality engineering actions
 
+- Listrak product payload fields (Category, Sub category, Meta 1, Meta 2, Meta 3) in the app settings to make it available for the admin to change.
+
 ## [2.0.2] - 2021-09-27
 
 ### Added